// BEEPTB_CUSTOMER_APP/src/app/shared/models/calculateAll.model.ts
export class CalculateAllModel {
    priceServiceDetailId: number;
    id: number;
    code: string;
    name: string;
    serviceId: number;
    vehicleTypeId: number;
    vehicleTypeName: string;
    priceFormulaId: number;
    dateFrom: any;
    dateTo: any;
    priority: any;
    value: any;
    valueFrom: any;
    valueTo: any;
    valuePlus: any;
    price: any;
}